refactor: Potencia renders from the block registry; drop SectionBlocks
Extracts the TDP state into a useTdp hook shared via PotenciaProvider by the
power-arc core, the GPU-clock card and the Auto‑TDP toggle. The core is registered
so a custom view can host it but stays fixed chrome in the Potencia tab; GPU/Auto
remain the reorderable blocks (rendered only when not stepped aside for another TDP
manager). With every section on SectionView, the old SectionBlocks + its markPresent
are removed.
